+++ src/main/java/algorithm/F.java
@@ -1,6 +1,6 @@
 package algorithm;
 public class F {
-	public Thing P1;
-	public Thing P2;
+	public Person P1;
+	public Person P2;
 	public long D;
 }
