+++ src/main/java/algorithm/Person.java
@@ -1,0 +1,22 @@
+package algorithm;
+
+import java.util.Date;
+
+public class Person {
+	public String name;
+	public Date birthDate;
+
+	public String getName() {
+		return name;
+	}
+	public void setName(String name) {
+		this.name = name;
+	}
+	public Date getBirthDate() {
+		return birthDate;
+	}
+	public void setBirthDate(Date birthDate) {
+		this.birthDate = birthDate;
+	}
+}
+
