--- src/main/java/algorithm/Thing.java
@@ -1,22 +1,0 @@
-package algorithm;
-
-import java.util.Date;
-
-public class Thing {
-	public String name;
-	public Date birthDate;
-
-	public String getName() {
-		return name;
-	}
-	public void setName(String name) {
-		this.name = name;
-	}
-	public Date getBirthDate() {
-		return birthDate;
-	}
-	public void setBirthDate(Date birthDate) {
-		this.birthDate = birthDate;
-	}
-}
-
