+++ src/test/java/test/FinderTests.java
@@ -11,14 +11,14 @@ import org.junit.Test;
 import algorithm.F;
 import algorithm.FT;
 import algorithm.Finder;
-import algorithm.Thing;
+import algorithm.Person;
 
 public class FinderTests {
 
-	Thing sue = new Thing();
-	Thing greg = new Thing();
-	Thing sarah = new Thing();
-	Thing mike = new Thing();
+	Person sue = new Person();
+	Person greg = new Person();
+	Person sarah = new Person();
+	Person mike = new Person();
 
 	@Before
 	public void setup() {
@@ -34,7 +34,7 @@ public class FinderTests {
 
 	@Test
 	public void Returns_Empty_Results_When_Given_Empty_List() {
-		List<Thing> list = new ArrayList<Thing>();
+		List<Person> list = new ArrayList<Person>();
 		Finder finder = new Finder(list);
 
 		F result = finder.Find(FT.One);
@@ -45,7 +45,7 @@ public class FinderTests {
 
 	@Test
 	public void Returns_Empty_Results_When_Given_One_Person() {
-		List<Thing> list = new ArrayList<Thing>();
+		List<Person> list = new ArrayList<Person>();
 		list.add(sue);
 
 		Finder finder = new Finder(list);
@@ -58,7 +58,7 @@ public class FinderTests {
 
 	@Test
 	public void Returns_Closest_Two_For_Two_People() {
-		List<Thing> list = new ArrayList<Thing>();
+		List<Person> list = new ArrayList<Person>();
 		list.add(sue);
 		list.add(greg);
 		Finder finder = new Finder(list);
@@ -71,7 +71,7 @@ public class FinderTests {
 
 	@Test
 	public void Returns_Furthest_Two_For_Two_People() {
-		List<Thing> list = new ArrayList<Thing>();
+		List<Person> list = new ArrayList<Person>();
 		list.add(mike);
 		list.add(greg);
 
@@ -85,7 +85,7 @@ public class FinderTests {
 
 	@Test
 	public void Returns_Furthest_Two_For_Four_People() {
-		List<Thing> list = new ArrayList<Thing>();
+		List<Person> list = new ArrayList<Person>();
 		list.add(sue);
 		list.add(sarah);
 		list.add(mike);
@@ -100,7 +100,7 @@ public class FinderTests {
 
 	@Test
 	public void Returns_Closest_Two_For_Four_People() {
-		List<Thing> list = new ArrayList<Thing>();
+		List<Person> list = new ArrayList<Person>();
 		list.add(sue);
 		list.add(sarah);
 		list.add(mike);
